correction du français (#494)
I have see that you have a french traduction but honestly sometimes i didn't understand the ideas behind the sentences thus i have decided to propose changes that i think are more understandable. I left some translations that are weird for me but may not be for a Canadian (in France we use a lot of English words unlike Canada which tries to translate everything into good French).