fix: general improvements to scheduled tasks
Telemetry and error reporting components are stripped, so I don't think that it's much of an issue not deleting some here.

I want to make this list pretty minimal and specific on what it deletes, as lots of these tasks won't ever run anyways by default unless needed for compatibility.

Additionally, deleting some folders currently errors with 'Access denied' in AME Wizard, so this will fix that.