fix(lockscreen): slideshow on boot
Removes the disabling of 'Use sign-in info to auto-finish setting up device after update or restart.'

Previously, there was a privacy concern was that mostly Discord would startup silently on the lockscreen without any control and auto-join a voice channel you were in, without you knowing. However, that has been fixed as Discord now asks you if you want to re-join instead of doing it for you.

Also fixes the lockscreen slideshow and widgets not working on boot.