fix: (dev) build script for Non-English environment (#1310)
### Questions
- [x] Did you test your changes or double-check that they work?
- [x] Did you read and follow the [Atlas Contribution
Guidelines](https://docs.atlasos.net/contributions/)?

### Describe your pull request

fix build script for Non-English environment.

I want to test my modified version of playbook, but the output of the
test target seems having a broken xml format.

It turns out that I use the Chinese environment and powershell use the
Chinese encoding to read and write `playbook.conf` file, that would
break the `Description` content which contains special characters.

So here is the fix, simply add `-Encoding` to the read and write
commands.